\name{RfmriVC-internal}

\alias{validConfig}
\alias{cHRF}
\alias{cHRFdp6}
\alias{cHRFdp3}
\alias{dgammadp6}
\alias{k.pen}
\alias{w.design}
\alias{wk}
\alias{w1}
\alias{x.design}
\alias{B.design}
\alias{E.design}
\alias{E.stern.design}
\alias{beta.hat.func}
\alias{beta.hat.ci}
\alias{beta.penkq.ci}
\alias{HRF.J.t.hat}
\alias{hrf.J.t.hat.penkq.ci}
\alias{hrf.J.t.hat.bayes.ci}
\alias{fc.gamma.k}
\alias{fc.sigma2}
\alias{fc.tau2.k}
\alias{fc.zeta}
\alias{VCM.gibbs}
\alias{nifti-class}
\alias{doFmriVC.Classical-method}
\alias{doFmriVC.Bayes-method}
\alias{doFmriVC-method}
\alias{betaHatNifti-method}
\alias{betaHat-methods}
\alias{setResultsVC-method}
\alias{resetResultsVC-method}
\alias{hrfHat-method}
\alias{hrfHatPoint-method}
\alias{beta.hat.plot}
\alias{calc.ImagePars}


\title{Internal RfmriVC functions}
\description{
  Internal RfmriVC functions.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
